Move all base64 logic into separate module
