+++ lib/remap-functions/src/base64_encoding/decode_base64.rs
@@ -1,4 +1,4 @@
-use crate::base64_util::Charset;
+use super::Charset;
 use remap::prelude::*;
 use std::str::FromStr;
